start refactoring some functions to a helper file
